-- new values on wire width menu
-- wire widths on old sketches are still valid (so no sketches need to be redone), but the old width values are no longer available

git-svn-id: http://fritzing.googlecode.com/svn/trunk@3491 85f94c43-f63a-0410-9d22-21439860c375

+++ fritzing/src/commands.cpp
@@ -683,7 +683,7 @@ QString WireColorChangeCommand::getParamString() const {
 
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-WireWidthChangeCommand::WireWidthChangeCommand(SketchWidget* sketchWidget, long wireId, int oldWidth, int newWidth, QUndoCommand *parent)
+WireWidthChangeCommand::WireWidthChangeCommand(SketchWidget* sketchWidget, long wireId, qreal oldWidth, qreal newWidth, QUndoCommand *parent)
 : BaseCommand(BaseCommand::SingleView, sketchWidget, parent)
 {
 	m_wireId = wireId;
