+++ fritzing/src/infographicsview.cpp
@@ -164,7 +164,7 @@ void InfoGraphicsView::initWire(Wire *, int penWidth) {
 }
 
 
-void InfoGraphicsView::getBendpointWidths(class Wire *, int w, int & w1, int & w2) {
+void InfoGraphicsView::getBendpointWidths(class Wire *, qreal w, qreal & w1, qreal & w2) {
 	Q_UNUSED(w);
 	Q_UNUSED(w1);
 	Q_UNUSED(w2);
