+++ fritzing/src/items/itembase.cpp
@@ -1353,42 +1353,42 @@ bool ItemBase::hasConnections()
 	return false;
 }
 
-void ItemBase::getConnectedColor(ConnectorItem *, QBrush * &brush, QPen * &pen, qreal & opacity, int & negativePenWidth) {
+void ItemBase::getConnectedColor(ConnectorItem *, QBrush * &brush, QPen * &pen, qreal & opacity, qreal & negativePenWidth) {
 	brush = &connectedBrush;
 	pen = &connectedPen;
 	opacity = 0.2;
 	negativePenWidth = 0;
 }
 
-void ItemBase::getNormalColor(ConnectorItem *, QBrush * &brush, QPen * &pen, qreal & opacity, int & negativePenWidth) {
+void ItemBase::getNormalColor(ConnectorItem *, QBrush * &brush, QPen * &pen, qreal & opacity, qreal & negativePenWidth) {
 	brush = &normalBrush;
 	pen = &normalPen;
 	opacity = normalConnectorOpacity;
 	negativePenWidth = 0;
 }
 
-void ItemBase::getUnconnectedColor(ConnectorItem *, QBrush * &brush, QPen * &pen, qreal & opacity, int & negativePenWidth) {
+void ItemBase::getUnconnectedColor(ConnectorItem *, QBrush * &brush, QPen * &pen, qreal & opacity, qreal & negativePenWidth) {
 	brush = &unconnectedBrush;
 	pen = &unconnectedPen;
 	opacity = 0.6;
 	negativePenWidth = 0;
 }
 
-void ItemBase::getChosenColor(ConnectorItem *, QBrush * &brush, QPen * &pen, qreal & opacity, int & negativePenWidth) {
+void ItemBase::getChosenColor(ConnectorItem *, QBrush * &brush, QPen * &pen, qreal & opacity, qreal & negativePenWidth) {
 	brush = &chosenBrush;
 	pen = &chosenPen;
 	opacity = normalConnectorOpacity;
 	negativePenWidth = 0;
 }
 
-void ItemBase::getHoverColor(ConnectorItem *, QBrush * &brush, QPen * &pen, qreal & opacity, int & negativePenWidth) {
+void ItemBase::getHoverColor(ConnectorItem *, QBrush * &brush, QPen * &pen, qreal & opacity, qreal & negativePenWidth) {
 	brush = &hoverBrush;
 	pen = &hoverPen;
 	opacity = normalConnectorOpacity;
 	negativePenWidth = 0;
 }
 
-void ItemBase::getEqualPotentialColor(ConnectorItem *, QBrush * &brush, QPen * &pen, qreal & opacity, int & negativePenWidth) {
+void ItemBase::getEqualPotentialColor(ConnectorItem *, QBrush * &brush, QPen * &pen, qreal & opacity, qreal & negativePenWidth) {
 	brush = &equalPotentialBrush;
 	pen = &equalPotentialPen;
 	opacity = 1.0;
